+++ app/view/constants.ts
@@ -1,0 +1,4 @@
+// Synchronize these with `constants.less`
+export const enum StyleConstants {
+    ScreenMediumMax = 800,
+}