+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Apply `-webkit-overflow-scrolling` on tables for inertia scrolling on iOS
+- Reset the visibility flag of the sidebar modal when it's hidden.
 
 ## [0.0.7] - 2018-12-31